test: move validation tests to separate suite
